+++ CleanWebApiTemplate.Host/Routes/Todo/Update/UpdateTodoRequest.cs
@@ -4,5 +4,5 @@ public struct UpdateTodoRequest
 {
     public string? Title { get; set; }
     public string? Description { get; set; }
-    public byte? Status { get; set; }
+    public int? Status { get; set; }
 }
